Reland "keysym look up with modifiers."
This reverts commit 4774f7437d3c676de009cd13c5c4f37f74c92449,
and made a fix on top of it.

The original CL reverts the following part.
- OnKeysym, starts looking up modifiers.

However, we got a regression with the change, because 1) ash-chrome
didn't send CapsLock modifier properly and 2) NumLock should be always
interpreted as if it is kept set on ChromeOS. As a result, keys with
these modifiers didn't work.
This CL also contains the fix for them, and workaround for the short
term backward compatibility, since lacros needs to run on older
ash-chrome.

BUG=1289236
TEST=Ran on DUT. Made sure Caps/NumLock works, too.

Change-Id: I03c008dc8365efef7ba5e928a08de135a99e6b1b
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3582437
Reviewed-by: Robert Kroeger <rjkroege@chromium.org>
Commit-Queue: Hidehiko Abe <hidehiko@chromium.org>
Cr-Commit-Position: refs/heads/main@{#991287}