[a11y] Re-enable HTMVideoElement a11y dump tests
While working on CL:3508289, I came to use the AccessibilityVideControls
test but it was disabled. Looking into it a bit more, I couldn't figure
out the reason behind the flakiness that got it disabled in the first
place (3+ years ago).

We are re-enabling these two HTMLVideoElement-related accessibility
dump tests in this CL because it's possible the flakiness got fixed in
the last few years, and if not, to get more data that will help us
hopefully fix the flakiness.

We are also disabling these tests for /blink only on macOS, since the
expectations are different over there. See crbug.com/1314896.

Bug: 916003,1314896
Change-Id: I52ca1fcc7046fcb80e530c2d2a5baad9ba4e07d2
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3576533
Reviewed-by: Daniel Libby <dlibby@microsoft.com>
Commit-Queue: Benjamin Beaudry <benjamin.beaudry@microsoft.com>
Cr-Commit-Position: refs/heads/main@{#991305}