suggest: allow assistive on mail.google.com/chat
This CL updates the url checking logic found in
`AssistiveSuggesterClientFilter` to inspect not only the domain of the
given url, but also the path. This allows us to support app urls where
we may want to allow an assistive feature on a specific path, but not
the root domain (ie. the current url for Chat, mail.google.com/chat).

BUG=b:228394142

Change-Id: Ieea0b9f8e9a1ac8289a707e9c69305e2771f7269
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3577763
Reviewed-by: Darren Shen <shend@chromium.org>
Commit-Queue: Curtis McMullan <curtismcmullan@chromium.org>
Cr-Commit-Position: refs/heads/main@{#991398}