RtpContributingSourceCache: Avoid theoretical use-after-free.
The `pc_` reference is a WeakPersistent, so it could become null causing
MaybeUpdateRtpSources() not to update the sources, resulting in
`rtp_sources` being null. This CL adds a guard to prevent
use-after-free.

Just a sanity check - I don't think this can happen in practise due to
RTCRtpReceiver having a strong reference to the PC which would prevent
the PC from going away while calling getContributingSources() is still
a possibility. But in case some assumption changes in the future...

Bug: chromium:1150311
Change-Id: If66cc8492f457c145f380adbbebaf1ab594974d7
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3578751
Reviewed-by: Harald Alvestrand <hta@chromium.org>
Commit-Queue: Henrik Boström <hbos@chromium.org>
Cr-Commit-Position: refs/heads/main@{#991405}