[iOS] Keep tableView items' index path in sync with the model
This CL is an attempt to fix a crash that reports invalid number of rows
in section when inserting a table cell. This is most likely the cause of
the tableView cells' indices and the tableViewModel items' index paths
are out of sync, which happens when a certain permission becomes not
accessible while the infobar modal is still visible.

Bug: 1314577
Change-Id: I2aa647b426a4b6005b3ba503fcfac858d5b0e53d
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3577421
Auto-Submit: Ginny Huang <ginnyhuang@chromium.org>
Reviewed-by: Gauthier Ambard <gambard@chromium.org>
Commit-Queue: Gauthier Ambard <gambard@chromium.org>
Cr-Commit-Position: refs/heads/main@{#991430}