[Android][UPM] Fix DCHECK triggered by sync callback
With https://crrev.com/c/3560442, the sync bridge tries to invoke a
forwarding method on the proxy instead of the callback of the store
directly.
Because the proxy binds itself weakly to that forwarding method, the
sequence checker will run into a DCHECK when the callback is called from
an incorrect thread.

Since the forwarding method doesn't need any information of the proxy
(aside for a DCHECK), the simplest fix is to move forwarding method out
of the proxy.

Bug: 1315523
Change-Id: Ia17aaf8c28c1073c135ef3ac7fd5a3ec372139bc
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3582201
Reviewed-by: Mohamed Amir Yosef <mamir@chromium.org>
Commit-Queue: Friedrich Horschig <fhorschig@chromium.org>
Cr-Commit-Position: refs/heads/main@{#991461}