[ios] Add context_menu provider API
Add a ContextMenuActionProvider function exposed by the provider API.
The internal implementation can implement it to insert menu elements
in the context menu.

The function is not used as the internal repository does not provide
an implementation of the function (and thus would not link if it were
used).

This is part of a multi-CL change:
 => https://crrev.com/c/3582202
  - https://crrev.com/i/4673653

Bug: 1290270
Change-Id: I85e29b2a86949089bcc24f0797d3f2afa7383719
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3582202
Reviewed-by: David Jean <djean@chromium.org>
Commit-Queue: Sylvain Defresne <sdefresne@chromium.org>
Cr-Commit-Position: refs/heads/main@{#991479}