Increase number of shards for ChromeOS MSAN bot
A recent update has caused an increase in run times and we need a
temporary increase in the number of shards used by the ChromeOS MSAN bot
in order to turn it green again. The number of shards can come back to
normal once we've fixed the cause of the slowdown.

Bug: chromium:1313907
Change-Id: I8d1bef217c5e2b87e1d420a93375ceda15a8cdbe
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3582639
Reviewed-by: Gary Tong <gatong@chromium.org>
Commit-Queue: Alexis Hétu <sugoi@google.com>
Cr-Commit-Position: refs/heads/main@{#991483}