Show rejected attribution triggers in separate table in internals UI
This makes the actual behavior of the API clearer, as prior to this CL
the internals UI made it appear that reports were actually generated,
albeit dropped, in the presence of attribution failures. This CL makes
it clearer by only showing reports that were ever stored or sent in
the Reports table, including ones that were stored but later dropped
in favor of higher-priority ones before being sent.

This also makes it possible to indicate when a trigger did not result
in a report due to no matching sources or no aggregatable histograms,
and will enable us to indicate when a trigger was rejected due to the
browser policy, which I will address in a followup CL.

Screenshot of changes in crbug.com/1295004#c7

Bug: 1295004, 1309834
Change-Id: I9260eb603616a660f0c24815482145ff5ac7fe89
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3575264
Reviewed-by: Nan Lin <linnan@chromium.org>
Reviewed-by: Robert Sesek <rsesek@chromium.org>
Commit-Queue: Andrew Paseltiner <apaseltiner@chromium.org>
Cr-Commit-Position: refs/heads/main@{#991484}