[AncestorChain] Add nested iframe case to WebLocks WPTs
To ensure that nested iframes with a cross-site ancestor
(such as the scenario A1->B->A2) are partitioned in the
WebLocks code, a new optional test case was added to the
existing WPTs. In this setup, when third-party storage
partitioning is enabled, both A1 and A2 can obtain the
same weblock, as they are in separate partitions.

Change-Id: Ie481db5fa5b5d9d5a2de118d6091c4d9afab0456
Fixed: 1312180
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3514221
Reviewed-by: Ben Kelly <wanderview@chromium.org>
Commit-Queue: Kyra Seevers <kyraseevers@chromium.org>
Cr-Commit-Position: refs/heads/main@{#991497}