Fix undefined float conversion.
The code which returned the image size for entropy calculations
incorrectly cast from int64 -> double -> uint64, which was not only
unnecessary, but caused UB when the original int was negative.

This removes the unnecessary float round trip, and clamps the output
to non-negative values.

Bug: 1315059
Change-Id: I6509b927546bfa1d03c04ac653acbc627d23cffb
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3579087
Reviewed-by: Yutaka Hirano <yhirano@chromium.org>
Commit-Queue: Ian Clelland <iclelland@chromium.org>
Cr-Commit-Position: refs/heads/main@{#991515}