Keep message center open when opened by keyboard from bubble launcher
If the user types a search into the bubble launcher, it switches to
the search page. If they type a keyboard shortcut like Alt-Shift-N the
notification center opens and takes focus. This also explicitly
dismisses the app list.

When the app list hide animation completes it clears the search box
text. This can cause a transition to the apps page, which can move
focus to the search box. The focus change causes the notification
center to close unexpectedly.

Suppress search box focus transitions when the app list is hiding.
The search box is focused on show, so there's no need to change its
focus during the hide.

Bug: 1313140
Test: added to ash_unittests
Change-Id: I1331305de753bf3df1a04a666f2dcc87388213e8
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3579673
Reviewed-by: Toni Barzic <tbarzic@chromium.org>
Commit-Queue: James Cook <jamescook@chromium.org>
Cr-Commit-Position: refs/heads/main@{#991541}