[Private Network Access] Never send preflight after forwarding headers.
Doing so otherwise might trigger crbug.com/1279376 again, if for some
reason a request fails with `kUnexpectedPrivateNetworkAccess` after
having already forwarded headers. This should not be the case thanks to
crrev.com/c/3552562 and crrev.com/c/3572052, but might happen again in
the future if maintainers forget to uphold the invariant.

In order to guard against that, we DCHECK that the scenario never
arises using NOTREACHED(). In production code, we simply fail the
request instead of crashing.

Bug: chromium:1279376
Change-Id: I97555036f6288641990b3cd1fc80834c139f03d2
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3582224
Reviewed-by: Maks Orlovich <morlovich@chromium.org>
Commit-Queue: Titouan Rigoudy <titouan@chromium.org>
Auto-Submit: Titouan Rigoudy <titouan@chromium.org>
Cr-Commit-Position: refs/heads/main@{#991583}