[Extension] Fix sendMessage emitting unhelpful errors for promises
This CL changes how errors are handled when message ports for
OneTimeMessageHandler listeners are closed before they are able to
reply, when they had indicated they intended to response
asynchronously. What this means for extensions is sendMessage() calls
will more accurately report errors for the receivers not returning an
expected delayed response vs not receiving an expected immediate
response. This also resolves a problem where calling sendMessage with
no callback argument would report an error if the listener didn't
respond, due to how expected responses were being treated for promise
based calls.

These changes are achieved by having the receiver's renderer send a
newly introduced IPC to the browser if one of the receivers indicated
that it intends to respond asynchronously. Then if the associated port
closes unexpectedly, an explicit error message describing this is sent
back to the opener's renderer.

Additionally the opener keeps track if the message port is being used
for a promise based call, as the default behavior for the port being
closed before a non-asynchronous response is received differs slightly
from the callback based case and does not need to report a default
error.

Tests have also been added to more thoroughly document and test these
behaviors.

Bug: 1304272
Change-Id: If507fdb364e6a1929defea3a6a8fcd3bb1cc4636
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3526376
Reviewed-by: Devlin Cronin <rdevlin.cronin@chromium.org>
Reviewed-by: Greg Kerr <kerrnel@chromium.org>
Commit-Queue: Tim <tjudkins@chromium.org>
Cr-Commit-Position: refs/heads/main@{#991650}