crates.py: cargo_pkg_name, cargo_pkg_description
These are two more standard environment variables set by cargo and
occasionally used by Rust crates (mostly to display help messages).
cargo_pkg_name is used by one of the new dependencies of autocxx.
cargo_pkg_description is not currently used, but as it's a standard
macro available in the 'clap' crate, we are likely to come across it
soon and so we may as well populate it now.

Bug: 1296155
Change-Id: Id2ef279ba339d77e6e265cc52ddbcc6660b43780
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3582689
Reviewed-by: danakj <danakj@chromium.org>
Commit-Queue: Adrian Taylor <adetaylor@chromium.org>
Cr-Commit-Position: refs/heads/main@{#991710}