Modify IntentChipButtonBrowserTest to remove apps.
This CL is used to prepare for CL:3580406. When the test finishes, the
app is not removed, and the preferred app setting is not cleared. This
might affect other browser tests.

This CL modifies IntentChipButtonBrowserTest to call UninstallWebApp to
remove apps saved in the ash-Chrome AppRegistryCache before existing the
tests to prevent affecting other tests.

BUG=1253250

Change-Id: Ibbf581fcfd46273cc583fb5eaa2284583b526ec3
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3572264
Reviewed-by: Tim Sergeant <tsergeant@chromium.org>
Reviewed-by: Scott Violet <sky@chromium.org>
Commit-Queue: Nancy Wang <nancylingwang@chromium.org>
Cr-Commit-Position: refs/heads/main@{#991729}