[FlexNG] Remove unnecessary TODO
I had added a TODO to update the intrinsic block size for columns
when the line extended beyond the last flex item. However, this can
only happen when there is a set flex block-size and the items are
aligned within the container.

As a result, this doesn't have any visible impact. I was going to add
a unit test to make sure that this was correct under the hood,
but it turns out the total_intrinsic_block_size for column flex
containers is based on the hypothetical main size of the columns
(which doesn't include alignment), so things are already incorrect
in this case. Thus, no need to try to get this correct for
fragmentation, and the TODO should be unnecessary.

Bug: 1315004
Change-Id: I1e8d5dfb5ded8c9baba94b972fb5790ad697859b
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3584157
Reviewed-by: Ian Kilpatrick <ikilpatrick@chromium.org>
Commit-Queue: Alison Maher <almaher@microsoft.com>
Cr-Commit-Position: refs/heads/main@{#991734}