[Extensions] Modify platformKeys API tests to support service workers.
This CL changes these tests to run from the extension's background
page, instead of from a tab, and adds parameterization to the tests
so they can be run as a legacy extension and a service worker-based
extension.

There are some interactive tests that should properly be run from
a tab, rather than from a background page, since they test an API
that opens a dialog box for user input. However, there is a delegate
in the tests that feeds input back instead of opening a dialog box,
so the current test coverage should be adequate. Also, isolating
those tests will require additional work, since there are
dependencies between the tests that will need to be fixed to move
the interactive tests.

Bug: 1303197
Change-Id: Ia0970f2e5e25eb68a7f414804eeb45243d98a727
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3560980
Reviewed-by: Devlin Cronin <rdevlin.cronin@chromium.org>
Commit-Queue: David Bertoni <dbertoni@chromium.org>
Cr-Commit-Position: refs/heads/main@{#991740}