[Downloads] Open CCT in the same task/window as parent activity
This CL fixes a multiwindow bug introduced in android T. When opening a
download, we were using the application context to fire the intent.
This causes the activity to be launched in a new task rather than having
affinity to the original window from which the download is being opened.
This CL uses the last focused activity to fix this issue.

TODO: Cleanup downloads code to pass the correct activity wherever
possible

Bug: 1313472
Change-Id: I27b25b9e9ce7f1429c830a01ef41d976f8f14100
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3583075
Reviewed-by: Theresa Sullivan <twellington@chromium.org>
Reviewed-by: Min Qin <qinmin@chromium.org>
Commit-Queue: Shakti Sahu <shaktisahu@chromium.org>
Cr-Commit-Position: refs/heads/main@{#991755}