Convert ios/web/find_in_page js_compile_bundle targets
Converts find in page JavaScript targets to be compiled with Rollup and
Terser instead of the closure compiler.

Additionally, removes the outer anonymous function which is now
redundant as the JavaScript produced by optimize_js always wraps the
resulting script in an anonymous function.

Bug: 1313654
Change-Id: I4a96ac32fae1b36972cd5cb4a35ef16edab0ca55
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3579995
Reviewed-by: Chris Lu <thegreenfrog@chromium.org>
Commit-Queue: Mike Dougherty <michaeldo@chromium.org>
Cr-Commit-Position: refs/heads/main@{#991770}