[segmentation] Move database ownership out of platform
Create a new class to own the databases and manage initialization and
maintenance in this class. No functional change in code in the cl.
Moves the pref storage to selector class which is the only one using
it. Removes all unnecessary includes from platform_service_impl.cc

Bug: 1307083
Change-Id: I19844252c6eadea6ec9d83da2c2e50a1c275afc2
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3580033
Reviewed-by: Min Qin <qinmin@chromium.org>
Commit-Queue: Siddhartha S <ssid@chromium.org>
Cr-Commit-Position: refs/heads/main@{#991785}