[Fenced Frames] Update client hints for fenced frames
For both MPArch and Shadow DOM fenced frames, we need to ensure
that we honor the permissions policy of the fenced frame in its
subtree when adding client hints.

With MPArch we can have nested frame trees, so when checking for
embedded frames, we need to ensure that we're not the outermost
rather than the main frame.

Thanks to both dtapuska@ and jbroman@ who provided a great deal
of assistance in writing these tests.

Bug: 1314749
Change-Id: I3834127217f8da6007c18efb75ccf5ce4da19b3a
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3577897
Reviewed-by: Dave Tapuska <dtapuska@chromium.org>
Reviewed-by: Shivani Sharma <shivanisha@chromium.org>
Reviewed-by: Alex Moshchuk <alexmos@chromium.org>
Commit-Queue: Ian Vollick <vollick@chromium.org>
Cr-Commit-Position: refs/heads/main@{#991787}