Change request_id used for WebAuthn to string type
To enable the WebAuthn in Lacros design, the request id has to be
changed to string type. Now Chrome OS always calls the new
ShowAuthDialogV2 API so we can change all usage of the request_id to
string type in Chrome. Also, the old API is not used anymore so we can
remove it and rename the new one back to ShowAuthDialog.

Bug: 1269528
Test: u2fd tast tests
Test: CQ
Change-Id: Ic3c6740d277bfd3d9e3d1af9f292a61a9a6c7af2
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3552721
Reviewed-by: Avi Drissman <avi@chromium.org>
Reviewed-by: Xiyuan Xia <xiyuan@chromium.org>
Reviewed-by: Martin Kreichgauer <martinkr@google.com>
Commit-Queue: Howard Yang <hcyang@google.com>
Cr-Commit-Position: refs/heads/main@{#991823}