WebUI TS: Add TS type definition for cr.ui
The files in this CL are used by Files app, this is a preparation to
start introducing TS to Files app.

The classes ContextMenuHandler, MenuButton and MenuItem  implement the
EventListener using `handleEvent(e: Event): void` which requires to
implement via interface EventListenerObject, thus the auto-generated
declaration doesn't work.

The classes AlertDialog and GridSelectionController fail to compile
because TS doesn't recognize the inheritance via __proto__ and doesn't
allow to use the `@override`.

List is used as a dependency of Grid, TS fails saying: "Cannot extend a
class List. Class constructor is marked as private."

Bug: b:228404944
Change-Id: I399f6caf85e7a72960cdd74d0f1decb4b0ed4eac
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3577655
Reviewed-by: Demetrios Papadopoulos <dpapad@chromium.org>
Reviewed-by: Luciano Pacheco <lucmult@chromium.org>
Commit-Queue: Luciano Pacheco <lucmult@chromium.org>
Cr-Commit-Position: refs/heads/main@{#991832}