Implement GPUQueueDescriptor, device/queue labels
Implements GPUQueueDescriptor as well as communicating label attribute
changes for GPUDevice/GPUQueue to Dawn. Labels from the descriptors are
not passed to Dawn at the time of requestDevice due to how the
descriptor is currently being passed, but is instead set using the
setLabel() method immediately after creation. This should generally not
be observable unless there are errors during device creation that would
have included the label.

Bug: dawn:1323
Change-Id: I27e608cb01ed76f9cdefd1740c3dcb370661409d
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3569265
Reviewed-by: Austin Eng <enga@chromium.org>
Commit-Queue: Brandon Jones <bajones@chromium.org>
Cr-Commit-Position: refs/heads/main@{#991837}