Desk Templates: Fix Null-dereference on shutdown
When shutting down, it is possible for a window activation to trigger.
In these cases, we end up trying to check the user session, even though
PrefServices may be null. We currently expected this to always be valid.

This CL changes it from a DCHECK to just exiting early in the cases that
PrefServices is null.

Bug: 1307138
Test: manual
Change-Id: I07d7f12207b1c2d1a2ad7dd6bfa86bb529c3642b
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3584803
Reviewed-by: Ahmed Fakhry <afakhry@chromium.org>
Commit-Queue: Richard Chui <richui@chromium.org>
Cr-Commit-Position: refs/heads/main@{#991843}